Merge Mining in Tradecraft without Forward Blocks

Release v13.2.4 adds support for so-called auxiliary proof-of-work (AKA "merge-mining") using bitcoin-compatible SHA256d, as a soft-fork.  Once activated, it allows miners to *simultaneously* work on both Bitcoin and Freicoin/Tradecraft blocks at the same time, finding proof-of-work solutions that satisfy one or the other, or both blocks.  It also adds poolside protection against block withholding attacks, allowing miners who opt into these new features to prevent their workers from withholding winning shares.

The only previously described soft-fork proof-of-work change is the Forward Blocks proposal.  This is a simplification of the ideas presented therein, which achieves the soft-fork proof-of-work change in a way that can be deployed more quickly, but is still compatible with future deployment of Forward Blocks.

The basic idea is that we require each new block to satisfy two proof of work challenges in series: first the auxiliary proof-of-work (merge-mining) challenge, which can be shared across multiple chains including bitcoin, then the existing (native) header challenge which specific to our chain.  Thus the order of work is that a miner first generates a Freicoin/Tradecraft block template, then commits to this block template in a special location within a bitcoin block.  The miner then sets about solving that block.  Each returned share is checked to see if it satisfies Tradecraft's auxiliary proof-of-work requirement.  When a winning share is found, the block template is modified to commit to the auxiliary proof-of-work, and it is returned to miners to have its native proof-of-work solved.

By having the auxiliary proof-of-work difficulty adjustment algorithm target a longer inter-block period, there will be a transition period during which the difficulty of the native header adjusts down and the auxiliary difficulty adjusts up.  During this period the overall proof-of-work function will not be progress-free, but the transition period will be short (a few months) and the benefits that come with merged mining well worth it.  At the end of the transition period, the native/compatibility headers will be at or near difficulty 1, which is effectively progress-free, or practically no limit at all once the protocol-cleanup fork activates.

The parent chain contains a commitment to the Tradecraft block at a fixed location.  The native chain likewise commits to the auxiliary proof-of-work in a fixed location as well.  The block header structure is modified to contain the information needed to validate these commitments.

The merge mining commitment in the parent chain is placed in the last 36 bytes of the last output of the last transaction.  This helps to minimize the size of the auxiliary header's Merkle proof.  The commitment format is a 32-byte commitment hash, a 4-byte commitment identifier, and the transaction's `nLockTime` field.  Note that no `lock_height` field is present, so Bitcoin can act as the parent chain, but the commitment identifier is chosen so that Tradecraft cannot.  This allows merge mining with Bitcoin and prevents Tradecraft blocks from containing such a commitment.

Note: since Bitcoin does not have consensus-enforced block-final transactions, and since every Bitcoin transaction requires an input, to merge-mine Tradecraft blocks a miner must have at least one input they control.

The commitment structure includes a hash of a pool-selected secret value, which is used for block-withholding protection.  The pool can allow auxiliary difficulty to be split across two challenges, one which is calculated in the usual way that is compatible with bitcoin, and the other using the block-withholding secret.  This serves two purposes: (1) a miner which submits a share cannot predict whether it will be a valid block or not, and therefore cannot grief a pool through block withholding, and (2) it decouples the finding of Tradecraft blocks from bitcoin blocks, since a share that meets the necessary work threshold for both chains isn't necessarily accepted by both.

Once merged mining is activated, every block has two proof-of-work targets: the native difficulty and the auxiliary (merge mined) difficulty.  And unlike previous merge-mining schemes like namecoin's, every block MUST have an auxiliary header: merge mining is not optional on a block-by-block basis.

Both the native and auxiliary difficulty adjustment algorithms separately track observed hash rate according to the block timestamps and their expected block arrival predictors, except that the auxiliary difficulty adjustment algorithm targets a longer inter-block interval: 15 minutes / 900 seconds, which results in 4 blocks per hour, or 96 blocks per day.

Once the auxiliary proof is satisfied, a proof then needs to be found for the native header.  The miner who finds this proof is allowed to change the block template coinbase's scriptSig or nSequence field, and/or the header's nNonce and nVersion fields.  Note that the native header's nTime field MUST be the same as the auxiliary header's, so nTime-rolling isn't permitted when solving the native header.

These changes are supported by the built-in stratum mining server, which will serve non-merge-mined work to connected miners in the default configuration.  That is to say, it constructs a fake "bitcoin" block for the auxiliary proof-of-work.  Work is ongoing to modify Bitcoin Core so as to support merge mining of Tradecraft blocks (no consensus changes are needed for bitcoin).

+++ src/crypto/hmac_sha256.h
@@ -41,7 +41,7 @@ public:
         return *this;
     }
     void Finalize(unsigned char hash[OUTPUT_SIZE]);
-    void Midstate(unsigned char hash[OUTPUT_SIZE*2], unsigned char* buffer, size_t* length);
+    void Midstate(unsigned char hash[OUTPUT_SIZE*2], unsigned char* buffer, uint64_t* length);
 };
 
 #endif // FREICOIN_CRYPTO_HMAC_SHA256_H
